move testing utils to testing module
