+++ src/executor/contract.rs
@@ -814,7 +814,7 @@ impl Drop for LockFile {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{starknet_stub::StubSyscallHandler, utils::test::load_starknet_contract};
+    use crate::{starknet_stub::StubSyscallHandler, utils::testing::load_starknet_contract};
     use cairo_lang_starknet_classes::contract_class::{
         version_id_from_serialized_sierra_program, ContractClass,
     };
