+++ src/libfuncs/bounded_int.rs
@@ -855,7 +855,7 @@ mod test {
 
     use crate::{
         context::NativeContext, execution_result::ExecutionResult, executor::JitNativeExecutor,
-        utils::test::load_cairo, OptLevel, Value,
+        utils::testing::load_cairo, OptLevel, Value,
     };
 
     #[test]
