+++ src/libfuncs/circuit.rs
@@ -1167,11 +1167,10 @@ fn build_u96_guarantee_verify<'ctx, 'this>(
 
 #[cfg(test)]
 mod test {
-
     use crate::{
         utils::{
             felt252_str,
-            test::{jit_enum, jit_panic, jit_struct, load_cairo, run_program_assert_output},
+            testing::{jit_enum, jit_panic, jit_struct, load_cairo, run_program_assert_output},
         },
         values::Value,
     };
