+++ src/libfuncs/ec.rs
@@ -507,7 +507,9 @@ pub fn build_zero<'ctx, 'this>(
 #[cfg(test)]
 mod test {
     use crate::{
-        utils::test::{jit_enum, jit_struct, load_cairo, run_program, run_program_assert_output},
+        utils::testing::{
+            jit_enum, jit_struct, load_cairo, run_program, run_program_assert_output,
+        },
         values::Value,
     };
     use cairo_lang_sierra::program::Program;
