+++ src/libfuncs/felt252.rs
@@ -329,7 +329,7 @@ pub fn build_is_zero<'ctx, 'this>(
 #[cfg(test)]
 pub mod test {
     use crate::{
-        utils::test::{load_cairo, run_program},
+        utils::testing::{load_cairo, run_program},
         values::Value,
     };
     use cairo_lang_sierra::program::Program;
