+++ src/libfuncs/felt252_dict.rs
@@ -183,7 +183,7 @@ pub fn build_squash<'ctx, 'this>(
 #[cfg(test)]
 mod test {
     use crate::{
-        utils::test::{
+        utils::testing::{
             jit_dict, jit_enum, jit_struct, load_cairo, run_program, run_program_assert_output,
         },
         values::Value,
