+++ src/libfuncs/felt252_dict_entry.rs
@@ -193,7 +193,7 @@ pub fn build_finalize<'ctx, 'this>(
 
 #[cfg(test)]
 mod test {
-    use crate::utils::test::{jit_dict, load_cairo, run_program, run_program_assert_output};
+    use crate::utils::testing::{jit_dict, load_cairo, run_program, run_program_assert_output};
 
     #[test]
     fn run_dict_insert() {
