+++ src/libfuncs/pedersen.rs
@@ -105,7 +105,7 @@ pub fn build_pedersen<'ctx>(
 
 #[cfg(test)]
 mod test {
-    use crate::utils::test::{load_cairo, run_program_assert_output};
+    use crate::utils::testing::{load_cairo, run_program_assert_output};
 
     use starknet_types_core::felt::Felt;
