+++ src/libfuncs/starknet.rs
@@ -3015,7 +3015,7 @@ pub fn build_get_class_hash_at<'ctx, 'this>(
 
 #[cfg(test)]
 mod test {
-    use crate::utils::test::{jit_enum, jit_struct, load_cairo, run_program_assert_output};
+    use crate::utils::testing::{jit_enum, jit_struct, load_cairo, run_program_assert_output};
     use cairo_lang_sierra::program::Program;
     use lazy_static::lazy_static;
     use starknet_types_core::felt::Felt;
