+++ src/types/nullable.rs
@@ -242,7 +242,7 @@ fn build_drop<'ctx>(
 #[cfg(test)]
 mod test {
     use crate::{
-        utils::test::{jit_enum, jit_struct, load_cairo, run_program},
+        utils::testing::{jit_enum, jit_struct, load_cairo, run_program},
         values::Value,
     };
     use pretty_assertions_sorted::assert_eq;
